requirements.txt
flask
redis
requests
bs4
mysql-connector-python